site/requirements.txt
streamlit
Pillow
streamlit-tags
google-cloud-storage